refactors the code
* creates square as a type of rectangle class instead of separate class.
* removes square test and add it into rectangle test.
* adds factory method into rectangle for creating rectangle and square

+++ src/com/tw/bootcamp/problem1/Shape.java
@@ -1,6 +1,6 @@
 package com.tw.bootcamp.problem1;
 
 public interface Shape {
-    int area();
-    int perimeter();
+    double area();
+    double perimeter();
 }
